Conflicts:
	WebContent/mypage.jsp
	WebContent/purchaseHistory.jsp

+++ src/com/internousdev/lilac/action/DeletePurchaseHistoryAction.java
@@ -1,6 +1,5 @@
 package com.internousdev.lilac.action;
 
-
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -11,14 +10,17 @@ import com.internousdev.lilac.dao.PurchaseHistoryInfoDAO;
 import com.internousdev.lilac.dto.PurchaseHistoryInfoDTO;
 import com.opensymphony.xwork2.ActionSupport;
 
-
-
 public class DeletePurchaseHistoryAction extends ActionSupport implements SessionAware{
 	private String categoryId;
 	private String sex;
 	private Map<String,Object>session;
 	public String execute(){
 		String result=ERROR;
+
+		if(session == null){
+			result = "timeout";
+			return result;
+		}
 
 		PurchaseHistoryInfoDAO purchaseHistoryInfoDAO=new PurchaseHistoryInfoDAO();
 		int count =purchaseHistoryInfoDAO.deleteAll(String.valueOf(session.get("loginId")));
